Added the leaderboard functionality, not connected to leaderboard screen yet
This is a fully working leaderboard generatetor and saver that saves to a text file. I couldn't get my text files to read and write in the project so created it on its own and imported it here so it should work with minor tweaks.

Also set game timer back to 5 minutes